+++ app/assets/stylesheets/components/_input.scss
@@ -76,11 +76,6 @@ input[type="checkbox"] {
   }
 }
 
-.select {
-  font-family: $font;
-  outline: none;
-}
-
 .dropzone {
   height: 300px;
   display: flex;
