+++ app/assets/stylesheets/partials/_fonts.scss
@@ -23,7 +23,7 @@ h1 {
 h2 {
 	font-weight: 500;
 	font-size: 24px;
-	margin-bottom: 20px;
+	margin-bottom: 10px;
 }
 
 h3, .work-title {
